+++ pkg/bookmark/exporting/service.go
@@ -5,6 +5,7 @@ package exporting
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/virtualtam/netscape-go/v2"
 	"github.com/virtualtam/sparklemuffin/pkg/bookmark"
@@ -22,23 +23,55 @@ func NewService(r Repository) *Service {
 	}
 }
 
+func (s *Service) getBookmarks(userUUID string, visibility Visibility) ([]bookmark.Bookmark, error) {
+	switch visibility {
+	case VisibilityAll:
+		return s.r.BookmarkGetAll(userUUID)
+	case VisibilityPrivate:
+		return s.r.BookmarkGetAllPrivate(userUUID)
+	case VisibilityPublic:
+		return s.r.BookmarkGetAllPublic(userUUID)
+	default:
+		return []bookmark.Bookmark{}, ErrVisibilityInvalid
+	}
+}
+
+// ExportAsJSONDocument exports a given user's bookmarks matching the
+// provided Visibility as a JSON bookmark document.
+func (s *Service) ExportAsJSONDocument(userUUID string, visibility Visibility) (*jsonDocument, error) {
+	bookmarks, err := s.getBookmarks(userUUID, visibility)
+	if err != nil {
+		return &jsonDocument{}, err
+	}
+
+	now := time.Now().UTC()
+
+	document := &jsonDocument{
+		Title:      fmt.Sprintf("SparkleMuffin export of %s bookmarks", visibility),
+		ExportedAt: now,
+	}
+
+	for _, b := range bookmarks {
+		jsonBookmark := jsonBookmark{
+			URL:         b.URL,
+			Title:       b.Title,
+			Description: b.Description,
+			Private:     b.Private,
+			Tags:        b.Tags,
+			CreatedAt:   b.CreatedAt,
+			UpdatedAt:   b.UpdatedAt,
+		}
+
+		document.Bookmarks = append(document.Bookmarks, jsonBookmark)
+	}
+
+	return document, nil
+}
+
 // ExportAsNetscapeDocument exports a given user's bookmarks matching the
 // provided Visibility as a Netscape bookmark document.
 func (s *Service) ExportAsNetscapeDocument(userUUID string, visibility Visibility) (*netscape.Document, error) {
-	var bookmarks []bookmark.Bookmark
-	var err error
-
-	switch visibility {
-	case VisibilityAll:
-		bookmarks, err = s.r.BookmarkGetAll(userUUID)
-	case VisibilityPrivate:
-		bookmarks, err = s.r.BookmarkGetAllPrivate(userUUID)
-	case VisibilityPublic:
-		bookmarks, err = s.r.BookmarkGetAllPublic(userUUID)
-	default:
-		err = ErrVisibilityInvalid
-	}
-
+	bookmarks, err := s.getBookmarks(userUUID, visibility)
 	if err != nil {
 		return &netscape.Document{}, err
 	}